# feat(db/redis): cache CPE#Titles