@shino 2024-06-25
Some measurement on the latency of GET CPE#Cache#Tiles: ~20msec by slowlog

```
% taskset -c 5 redis-server
% for i in {1..50}; do taskset -c 1 redis-cli get CPE#Cache#Titles > /dev/null ; done

% redis-cli --csv slowlog get 50 | sed 's/"",/\n/g' | grep 'CPE#Cache#Titles' | cut -d, -f3 | histogram.py -b 30 -m 10000 -x 90000
# NumSamples = 50; Min = 10000.00; Max = 90000.00
# Mean = 20371.620000; Variance = 92915706.475600; SD = 9639.279355; Median 18815.000000
# each ∎ represents a count of 1
10000.0000 - 12666.6667 [     0]:
12666.6667 - 15333.3333 [     0]:
15333.3333 - 18000.0000 [    14]: ∎∎∎∎∎∎∎∎∎∎∎∎∎∎
18000.0000 - 20666.6667 [    30]: ∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
20666.6667 - 23333.3333 [     4]: ∎∎∎∎
23333.3333 - 26000.0000 [     1]: ∎
26000.0000 - 28666.6667 [     0]:
28666.6667 - 31333.3333 [     0]:
31333.3333 - 34000.0000 [     0]:
34000.0000 - 36666.6667 [     0]:
36666.6667 - 39333.3333 [     0]:
39333.3333 - 42000.0000 [     0]:
42000.0000 - 44666.6667 [     0]:
44666.6667 - 47333.3333 [     0]:
47333.3333 - 50000.0000 [     0]:
50000.0000 - 52666.6667 [     0]:
52666.6667 - 55333.3333 [     0]:
55333.3333 - 58000.0000 [     0]:
58000.0000 - 60666.6667 [     0]:
60666.6667 - 63333.3333 [     0]:
63333.3333 - 66000.0000 [     0]:
66000.0000 - 68666.6667 [     0]:
68666.6667 - 71333.3333 [     0]:
71333.3333 - 74000.0000 [     0]:
74000.0000 - 76666.6667 [     0]:
76666.6667 - 79333.3333 [     0]:
79333.3333 - 82000.0000 [     0]:
82000.0000 - 84666.6667 [     0]:
84666.6667 - 87333.3333 [     1]: ∎
87333.3333 - 90000.0000 [     0]:
```